Now with locally cached Methods!
Added functionality for the script to check for a locally
cached json file that the user can select on initialization.

If one isn't passed, and/or it's not found, it creates one at
./Methods.json

It then uses the locally cached Methods list for all interaction
thereby saving itself an API call on initialization as compared with
previous version.